# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chivra AI Social Network API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Contact {
  id: number;
  name: string;
  gender: string;
  personalityTone: string;
  languageStyle: string;
  emotionalBehavior: string;
  /** @nullable */
  avatarUrl?: string | null;
  activityState: string;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface ContactSummary {
  id: number;
  name: string;
  gender: string;
  /** @nullable */
  avatarUrl?: string | null;
  activityState: string;
  relationshipState: string;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
}

export interface CreateContactBody {
  name: string;
  gender: string;
  personalityTone: string;
  languageStyle: string;
  emotionalBehavior: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface UpdateContactBody {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
  activityState?: string;
  /** @nullable */
  bio?: string | null;
}

export interface ContactActivity {
  id: number;
  activityState: string;
  /** @nullable */
  lastSeenAt?: string | null;
}

export interface Message {
  id: number;
  contactId: number;
  sender: string;
  content: string;
  messageType: string;
  isRead: boolean;
  createdAt: string;
}

export interface SendMessageBody {
  content: string;
  messageType?: string;
}

export interface MarkReadResponse {
  updated: number;
}

export interface Relationship {
  contactId: number;
  state: string;
  updatedAt: string;
}

export interface UpdateRelationshipBody {
  state: string;
}

export interface Memory {
  contactId: number;
  /** @nullable */
  userName?: string | null;
  facts: string[];
  /** @nullable */
  emotionalContext?: string | null;
  updatedAt: string;
}

export interface StatusPost {
  id: number;
  contactId: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface StatusPostWithContact {
  id: number;
  contactId: number;
  contactName: string;
  /** @nullable */
  contactAvatarUrl?: string | null;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  totalContacts: number;
  totalUnread: number;
  activeRelationships: number;
  friendCount: number;
  partnerCount: number;
  onlineCount: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface SendOpenaiVoiceMessageBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export type ListMessagesParams = {
  limit?: number;
  before?: number;
};