# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chivra AI Social Network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all AI contacts with latest message and unread count
      responses:
        "200":
          description: List of AI contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSummary"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a new AI contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a single AI contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update an AI contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete an AI contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /contacts/{id}/activity:
    get:
      operationId: getContactActivity
      tags: [contacts]
      summary: Get contact activity state and online status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactActivity"
  /contacts/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a chat with an AI contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a user message and get an AI streaming response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: SSE stream of AI reply chunks
          content:
            text/event-stream: {}
  /contacts/{id}/messages/mark-read:
    post:
      operationId: markMessagesRead
      tags: [messages]
      summary: Mark all messages in a chat as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkReadResponse"
...
  /contacts/{id}/memory:
    get:
      operationId: getMemory
      tags: [contacts]
      summary: Get the AI memory context for a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Memory data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gender:
          type: string
        personalityTone:
          type: string
        languageStyle:
          type: string
        emotionalBehavior:
          type: string
        avatarUrl:
          type: ["string", "null"]
        activityState:
          type: string
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - gender
        - personalityTone
        - languageStyle
        - emotionalBehavior
        - activityState
        - createdAt
    ContactSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gender:
          type: string
        avatarUrl:
          type: ["string", "null"]
        activityState:
          type: string
        relationshipState:
          type: string
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
        unreadCount:
          type: integer
      required:
        - id
        - name
        - gender
        - activityState
        - relationshipState
        - unreadCount
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        gender:
          type: string
        personalityTone:
          type: string
        languageStyle:
          type: string
        emotionalBehavior:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
      required:
        - name
        - gender
        - personalityTone
        - languageStyle
        - emotionalBehavior
    UpdateContactBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        activityState:
          type: string
        bio:
          type: ["string", "null"]
    ContactActivity:
      type: object
      properties:
        id:
          type: integer
        activityState:
          type: string
        lastSeenAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - activityState
    Message:
      type: object
      properties:
        id:
          type: integer
        contactId:
          type: integer
        sender:
          type: string
        content:
          type: string
        messageType:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contactId
        - sender
        - content
        - messageType
        - isRead
        - createdAt
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        messageType:
          type: string
      required:
        - content
    MarkReadResponse:
      type: object
      properties:
        updated:
          type: integer
      required:
        - updated
...
    Memory:
      type: object
      properties:
        contactId:
          type: integer
        userName:
          type: ["string", "null"]
        facts:
          type: array
          items:
            type: string
        emotionalContext:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - contactId
        - facts
        - updatedAt
...