# lib/api-zod/src/generated/types/contactSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chivra AI Social Network API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSummary {
  id: number;
  name: string;
  gender: string;
  /** @nullable */
  avatarUrl?: string | null;
  activityState: string;
  relationshipState: string;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: Date | null;
  unreadCount: number;
}