# lib/api-zod/src/generated/types/createContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chivra AI Social Network API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactBody {
  name: string;
  gender: string;
  personalityTone: string;
  languageStyle: string;
  emotionalBehavior: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}