# lib/api-zod/src/generated/types/updateContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chivra AI Social Network API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateContactBody {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
  activityState?: string;
  /** @nullable */
  bio?: string | null;
}